Refactor the env checker
